requirements.txt
streamlit
openai
python-dotenv
pillow
requests
aiohttp